requirements.txt
selenium==3.14.0
Appium-Python-Client==0.28
xmlrunner==1.7.7
requests==2.19.1